pyproject.toml
[project]
name = "model-deprecation-tracker"
version = "1.0.0"
description = "Tracks AI model deprecation schedules across providers"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "icalendar>=5.0",
    "python-dateutil>=2.8",
    "orjson>=3.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]
